Refactor auth and user separatedly
