+++ src/controllers/userController.js
@@ -1,0 +1,26 @@
+const { validationResult } = require('express-validator');
+const UsersService = require('../services/users')
+
+/*
+  Register new user
+*/
+const add = async (req, res, next) => {
+  try {
+    const errors = validationResult(req)
+
+    if (!errors.isEmpty()) {
+      res.status(400).send("Invalid object!")
+      return
+    }
+
+    const usersService = new UsersService()
+    const message = await usersService.add(req.body)
+    res.send(message)
+  } catch(error) {
+    next(error)
+  }
+}
+
+module.exports = {
+  add,
+}
