+++ src/utils/constants.js
@@ -1,4 +1,5 @@
-class constants {
-}
+const DAY_IN_MILLISECONDS = 24 * 60 * 60 * 1000;
 
-module.exports = constants
+module.exports = {
+  DAY_IN_MILLISECONDS
+}
